check.h hotfix #3

+++ decoder/inc/check.h
@@ -44,11 +44,11 @@ const uint32_t CRC_COMPUTE(uint32_t* memory) {
 }
 
 void CHECKER_REMEMBER_CHANNEL(uint32_t* channel_location) {
-    CRC_RESULTS[(channel_location / CHANNEL_SIZE) % CHANNELS] = CRC_COMPUTE(channel_location);
+    CRC_RESULTS[((uint32_t)channel_location / CHANNEL_SIZE) % CHANNELS] = CRC_COMPUTE(channel_location);
 }
 
 const bool CHECKER_VERIFY_CHANNEL(uint32_t* channel_location) {
-    return CRC_COMPUTE(channel_location) == CRC_RESULTS[(channel_location / CHANNEL_SIZE) % CHANNELS];
+    return CRC_COMPUTE(channel_location) == CRC_RESULTS[((uint32_t)channel_location / CHANNEL_SIZE) % CHANNELS];
 }
 
 void CHECKER_END() {
